refactor: redux feature 폴더 접근법 적용
