+++ src/hooks/useAppDispatch.ts
@@ -1,6 +1,6 @@
 import { useDispatch } from 'react-redux';
 import type { Action } from 'redux';
-import { RootState } from 'redux/reducers';
+import { RootState } from 'redux/rootReducer';
 import type { ThunkDispatch } from 'redux-thunk';
 
 type AppDispatch<T extends Action> = ThunkDispatch<RootState, null, T>;
