+++ src/redux/itemList/itemListThunk.ts
@@ -2,7 +2,7 @@ import { BASE_URL } from 'apis';
 import axios from 'axios';
 import type { Dispatch } from 'redux';
 
-import { ItemListAction, ItemListActionType } from '../actions/itemList';
+import { ItemListAction, ItemListActionType } from './itemList';
 
 export const getItemList = () => async (dispatch: Dispatch<ItemListAction>) => {
   dispatch({ type: ItemListActionType.GET_ITEM_LIST_START });
